real_requirements.txt
psycopg2
django
djangorestframework
python-dotenv
pandas
prospector
django-cors-headers
django-import-export
django-extensions
jinja2